Starting with build automation
When starting an open source project you want to have a proper build setup that can be automatically building the sources on each commit. This is where Travis-CI comes in. All it needs is a file that tells it what the environment looks like and off we go. So here's the config file for it.