First try of an implementation
An octree is an ordered array of child descriptors and data. The
descriptors tell the octree how its topology is looking like and the
data is stored at specified places to be accessed real quick.

So far the basics of a child descriptor is implemented, I need to get my
head around the data alignment as the nvidia implementation suggests.