Defining guards
Before return a node at a certain position, we have to find out if this
node exists and if the parameters for the function were correct. If the
element requested exists, it has to be returned, which is still missing
in this commit. This is because of a problem with the data structure I
currently have. C++ can increment a pointer to iterate through the child
nodes. In C# I presumably have to use a generic list which uses an array
as backing data structure.