Restructuring the project
The former intend of this project was to create a library that is providing generic octrees. But the implementation of a generic octree is without application when no one is putting data in it. Thus this repository is changed to build a graphics platform that is able to render octrees. In the upcoming commits, there will be implementations of renderers and the generic octree will become a voxel octree as seen in various scientific publications. The intend still is to get an octree structure for C# that is fast and efficient.